backend/requirements.txt
absl-py==2.0.0
astunparse==1.6.3
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssmin==0.2.0
Flask==3.0.0
Flask-Assets==2.1.0
Flask-Cors==4.0.0
Flask-Minify==0.42
flatbuffers==23.5.26
gast==0.5.4
gevent==23.9.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
greenlet==3.0.1
grpcio==1.59.2
h5py==3.10.0
htmlmin==0.1.12
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsmin==3.0.1
keras==2.15.0
lesscpy==0.15.1
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
ml-dtypes==0.2.0
numpy==1.26.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
Pillow==10.1.0
ply==3.11
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyExifTool==0.5.6
python-configuration==0.9.1
python-dotenv==1.0.0
rcssmin==1.1.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
typing_extensions==4.8.0
urllib3==2.1.0
webassets==2.0
Werkzeug==3.0.1
wrapt==1.14.1
xxhash==3.4.1
zope.event==5.0
zope.interface==6.1
